makes requested PR changes
