+++ src/actions/actionTypes.js
@@ -17,9 +17,8 @@ export const FETCH_RENT_REQUEST = 'FETCH_RENT_REQUEST';
 export const DELETE_RENT_REQUEST = 'DELETE_RENT_REQUEST';
 export const EDIT_RENT_REQUEST = 'EDIT_RENT_REQUEST';
 
-// CLUBS
+//Clubs
 export const FETCH_CLUBS = 'FETCH_CLUBS';
-
 
 //News
 export const CREATE_NEWS = 'CREATE_NEWS';
@@ -45,5 +44,3 @@ export const MODAL_OPEN = 'MODAL_OPEN';
 export const MODAL_CLOSE = 'MODAL_CLOSE';
 
 export const GET_PAGES = 'GET_PAGES';
-
-
