+++ src/actions/index.js
@@ -1,14 +1,25 @@
+import { useDispatch } from 'react-redux';
 import axios from 'axios';
-
-import { GET_PAGES, MODAL_OPEN, MODAL_CLOSE } from './actionTypes';
+import { GET_PAGES, FETCH_CLUBS, MODAL_OPEN, MODAL_CLOSE } from './actionTypes';
 
 const api = 'https://dev.huoa.org/wp-json';
 
-export const getPages = () => async (dispatch) => {
-  const response = await axios.get(`${api}/wp/v2/pages?page=1&per_page=100`);
-  const data = response.data.reduce((allData, { slug, title, acf }) => ({ ...allData, [slug]: { title: title.rendered, ...acf }}), {});
+export const Initialize = () => { 
 
-  dispatch({ type: GET_PAGES, payload: data });
+  const getPages = () => async (dispatch) => {
+    const response = await axios.get(`${api}/wp/v2/pages?page=1&per_page=100`);
+    const data = response.data.reduce((allData, { slug, title, acf }) => ({ ...allData, [slug]: { title: title.rendered, ...acf }}), {});
+    dispatch({ type: GET_PAGES, payload: data });
+  };
+  
+  const getClubs = () => async (dispatch) => {
+    const response = await axios.get(`${api}/wp/v2/clubs`);
+    dispatch({ type: FETCH_CLUBS, payload: response.data });
+  };
+  
+  const dispatch = useDispatch();
+  dispatch(getClubs());
+  dispatch(getPages());
 };
 
 export const openModal = (modalContent) => {
